Key Placeholders
New folder for placeholder art.
Put a sprite by the x/y key counter in the corner.
Changed one of the key pickups to look like a key.
Hid the other key pickup. Did you know there were two?
Changed the size of the TimerText just to see if I could fix the
blurriness the text seems to has on device (no luck).