This is where I had to stop today....
Ugh. I'm sorry.

So in GameLogic, PlayWinCondition, I still need to actually call the UI
object's ShowPanels stuff for the EndGamePanel.

I created an object that ShowPanels references but its mostly a copy of
main menu. I need to write it a new script and create the end game
options.

Notes:
-GateLogic, about Line 35. If the way I call the sound is too weird,
tell me and I will change it later. This was less work and I wanted to
get this done so I could finish the endgame stuff.